+++ CHANGELOG.md
@@ -3,6 +3,7 @@ vNext
 **New features or improvements**
 - Allow [[..]] filter notation in all text panels (markdown/html/text) (Issue #511)
 - New legend display option "Align as table" (Issue #136)
+- New legend display option "Right side", will show legend to the right of the graph (Issue #556)
 - Enhanced InfluxDB series aliasing (legend names) with pattern replacements (Issue #525)
 
 **Changes**